+++ tests/async_tests/test_project.py
@@ -4,10 +4,8 @@ import httpx
 from munch import Munch
 from asyncmock import AsyncMock
 
-from tests.utils.test_util import (
-    async_client as client,  # noqa: F401
-    project_body,
-)
+from tests.utils.util import project_body
+from tests.utils.fixtures import async_client as client  # noqa: F401
 
 
 @pytest.mark.asyncio
