+++ tests/sync_tests/test_config.py
@@ -1,6 +1,6 @@
 import json
 from rapyuta_io_sdk_v2.config import Configuration
-from tests.utils.test_util import mock_config  # noqa: F401
+from tests.utils.util import mock_config  # noqa: F401
 
 
 def test_from_file(mocker):
