+++ tests/sync_tests/test_configtree.py
@@ -3,7 +3,8 @@ import pytest
 from munch import Munch
 from pytest_mock import MockerFixture
 
-from tests.utils.test_util import client, configtree_body  # noqa: F401
+from tests.utils.util import configtree_body  # noqa: F401
+from tests.utils.fixtures import client  # noqa: F401
 
 
 def test_list_configtrees_success(client, mocker: MockerFixture):  # noqa: F811
