+++ tests/sync_tests/test_main.py
@@ -2,7 +2,7 @@ import httpx
 import pytest
 from pytest_mock import MockerFixture
 
-from tests.utils.test_util import client  # noqa: F401
+from tests.utils.fixtures import client  # noqa: F401
 
 
 def test_get_auth_token_success(client, mocker: MockerFixture):  # noqa: F811
